==> 3_merge.ipynb <==
import pandas as pd
import geopandas as gpd
from dask.distributed import Client
import dask.dataframe as dd
from dask_jobqueue import SLURMCluster
import os
import matplotlib.pyplot as plt
import constants
import numpy as np
from panel_utilities import get_value_variable_names
import dask_geopandas
INPUT_DATA_EVICTIONS = "../data/02_intermediate/evictions.csv"
INPUT_DATA_TRACTS = "../data/02_intermediate/tracts.csv"
INPUT_DATA_TAX_PARCELS = "../data/02_intermediate/tax_parcels.gpkg"
INPUT_DATA_ZESTIMATES = "../data/02_intermediate/zestimates.csv"
INPUT_DATA_CRIME = "../data/01_raw/crime_incidents"
INPUT_DATA_NEIGHBORHOODS = "../data/01_raw/Boston_Neighborhoods/Boston_Neighborhoods.shp"
OUTPUT_DATA_ZILLOW = "../data/03_cleaned/zestimates_analysis.csv"
OUTPUT_DATA_CRIME = "../data/03_cleaned/crime_analysis_monthly.csv"
OUTPUT_DATA_NEIGHBORHOOD_CRIME_COUNTS = "../data/03_cleaned/neighborhood_crime_counts.csv"
OUTPUT_TABLES = "../output/final_paper/tables/"
VERBOSE = True
N_PARTITIONS = 1
value_vars_to_concat = []  # A list of DataFrames, where each DataFrame contains the panel data for a single outcome variable and has case_number as its index.

## 1. Loading Evictions Data

# Load evictions data.
with open(INPUT_DATA_EVICTIONS, 'r') as file:
    all_column_names = set(file.readline().replace("\"", "").replace("\n", "").split(","))
to_drop = {'Accuracy Score', 'Accuracy Type', 'Number', 'Street', 'Unit Type', 'Unit Number',
           'State', 'Zip', 'Country', 'Source', 'Census Year', 'State FIPS', 'County FIPS',
           'Place Name', 'Place FIPS', 'Census Tract Code', 'Census Block Code', 'Census Block Group',
           'Metro/Micro Statistical Area Code', 'Metro/Micro Statistical Area Type',
           'Combined Statistical Area Code', 'Metropolitan Division Area Code', 'court_location',
           'defendant', 'defendant_atty', 'defendant_atty_address_apt',
           'defendant_atty_address_city', 'defendant_atty_address_name', 'defendant_atty_address_state',
           'defendant_atty_address_street', 'defendant_atty_address_zip', 'docket_history', 'judgment_for',
           'judgment_total', 'plaintiff', 'plaintiff_atty', 'plaintiff_atty_address_apt',
           'plaintiff_atty_address_city', 'plaintiff_atty_address_name', 'plaintiff_atty_address_state',
           'plaintiff_atty_address_street', 'plaintiff_atty_address_zip', 'Metropolitan Division Area Name',
           'property_address_city', 'property_address_state', 'property_address_street',
           'property_address_zip'}
evictions_df = pd.read_csv(INPUT_DATA_EVICTIONS, usecols=set(all_column_names) - set(to_drop))
sample_restriction_table_index = []
sample_restriction_table_values = []

# Restrict to cases in Boston.
boston_mask = ((evictions_df['County'] == "Suffolk County") & (~evictions_df['City'].isin(["Chelsea", "Revere", "Winthrop"])))
if VERBOSE:
    print(f"Restricting to {boston_mask.sum()} observations which are from Boston.")
evictions_df = evictions_df.loc[boston_mask, :]
original_N = len(evictions_df)
if VERBOSE:
    print(f"Beginning with {original_N} observations.")
sample_restriction_table_index.append("Case Filed in Boston")
sample_restriction_table_values.append(original_N)



# Drop cases where initiating action is unknown
mask = evictions_df['initiating_action'] == "Summary Process - Residential (c239)"
if VERBOSE:
    print(f"Dropping {mask.sum()} cases for which the initiating action could not be scraped")
evictions_df = evictions_df.loc[~mask, :]
sample_restriction_table_index.append("Non-missing case initiating action")
sample_restriction_table_values.append(len(evictions_df))

# Drop cases initiated due to foreclosure
mask = evictions_df['initiating_action'].str.contains("Foreclosure")
if VERBOSE:
    print(f"Dropping {mask.sum()} cases which began due to foreclosure")
evictions_df = evictions_df.loc[~mask, :]
sample_restriction_table_index.append("Case initiated for reason other than foreclosure")
sample_restriction_table_values.append(len(evictions_df))



# Drop cases missing latest_docket_date.
mask = evictions_df['latest_docket_date'].notna()
if VERBOSE:
    print(
        f"Dropping {evictions_df['latest_docket_date'].isna().sum()} observations where latest_docket_date is missing.")
evictions_df = evictions_df.loc[mask, :]
sample_restriction_table_index.append("Non-missing latest docket date")
sample_restriction_table_values.append(len(evictions_df))

# Drop malformed addresses.
if VERBOSE:
    print(f"Dropping {evictions_df['property_address_full'].str.contains('span, span span').sum()} observations which "
          f"have malformed addresses.")
evictions_df = evictions_df.loc[~evictions_df['property_address_full'].str.contains("span, span span"), :]

# Drop missing addresses.
no_address_info_mask = (evictions_df['property_address_full'].isna())
if VERBOSE:
    print(
        f"Dropping {no_address_info_mask.sum()} rows missing property_address_full")
evictions_df = evictions_df.loc[~no_address_info_mask, :]
sample_restriction_table_index.append("Non-missing property address")
sample_restriction_table_values.append(len(evictions_df))

# Add file month and year to dataset.
evictions_df.loc[:, 'file_month'] = pd.to_datetime(evictions_df['file_date']).dt.strftime('%Y-%m')
evictions_df.loc[:, 'file_year'] = pd.to_datetime(evictions_df['file_date']).dt.year

# Add latest docket month and year to dataset.
evictions_df.loc[:, 'latest_docket_month'] = pd.to_datetime(evictions_df['latest_docket_date']).dt.strftime('%Y-%m')
evictions_df.loc[:, 'latest_docket_year'] = pd.to_datetime(evictions_df['latest_docket_date']).dt.year

# Drop cases where disposition found is other.
disposition_found_other_mask = (evictions_df['disposition_found'] == "Other") | (evictions_df['disposition_found'].isna())
if VERBOSE:
    print(f"Dropping {disposition_found_other_mask.sum()} cases where disposition_found is \"Other\" or missing.")
evictions_df = evictions_df.loc[~disposition_found_other_mask, :]
sample_restriction_table_index.append("Cases for which disposition could be scraped")
sample_restriction_table_values.append(len(evictions_df))

# Drop cases resolved via mediation.
mediated_mask = evictions_df['disposition_found'] == "Mediated"
if VERBOSE:
    print(f"Dropping {mediated_mask.sum()} cases resolved through mediation.")
evictions_df = evictions_df.loc[~mediated_mask, :]
sample_restriction_table_index.append("Case not resolved through mediation")
sample_restriction_table_values.append(len(evictions_df))

# Drop cases concluded in April 2020 or later.
pre_pandemic_months = ['2019-04',
 '2019-05',
 '2019-06',
 '2019-07',
 '2019-08',
 '2019-09',
 '2019-10',
 '2019-11',
 '2019-12',
 '2020-01',
 '2020-02',
 '2020-03']
pre_pandemic_mask = evictions_df['latest_docket_month'].isin(pre_pandemic_months)
evictions_df = evictions_df.loc[pre_pandemic_mask, :]
if VERBOSE:
    print(f"Dropping {(~pre_pandemic_mask).sum()} cases which concluded after pandemic began")
sample_restriction_table_index.append("Case concluded before April 2020")
sample_restriction_table_values.append(len(evictions_df))

# Drop rows which contain inconsistent values of disposition_found and judgment_for_pdu.
# First, we drop cases where disposition_found is "Defaulted" but judgment_for_pdu is "Defendant"
inconsistent_mask_1 = ((evictions_df['disposition_found'] == "Defaulted") & (evictions_df['judgment_for_pdu'] == "Defendant"))
if VERBOSE:
    print(f"Dropping {inconsistent_mask_1.sum()} observations where disposition_found is \"Defaulted\" but judgment_for_pdu is \"Defendant\".")
evictions_df = evictions_df.loc[~inconsistent_mask_1, :]
          
# Clean the values in the judgment_for_pdu variable.
judgment_for_pdu_replacement_dict = {"unknown": "Unknown",
                                     "plaintiff": "Plaintiff",
                                     "defendant": "Defendant"}
evictions_df.loc[:, "judgment_for_pdu"] = (evictions_df.loc[:, "judgment_for_pdu"]
                                           .replace(judgment_for_pdu_replacement_dict))

# Replace missing values in money judgment column with zeroes.
evictions_df.loc[:, 'judgment'] = evictions_df['judgment'].fillna(0)

# Rename duration to case_duration.
evictions_df = evictions_df.rename(columns={'duration': 'case_duration'})

# Build sample restriction table
sample_restriction_table = pd.DataFrame()
sample_restriction_table["Restriction"] = sample_restriction_table_index
sample_restriction_table["Observations"] = sample_restriction_table_values                         
sample_restriction_table = sample_restriction_table.set_index("Restriction")

# Export to LaTeX.
filename = os.path.join(OUTPUT_TABLES, "sample_restriction.tex")
sample_restriction_table.style.format(formatter="{:,.0f}").to_latex(filename, hrules=True)
sample_restriction_table

## 2. Merging Evictions With Census Tract Characteristics

# Merge with census tract characteristics.
evictions_df = evictions_df.rename(columns={'Full FIPS (tract)': 'tract_geoid'})
evictions_tracts_df = evictions_df.merge(pd.read_csv(INPUT_DATA_TRACTS, dtype={'tract_geoid': float}),
                                  on='tract_geoid',
                                  how='left',
                                  validate='m:1').set_index('case_number')
if VERBOSE:
    print(f"Successfully merged {evictions_tracts_df['med_hhinc2016'].notna().sum()} observations with census tracts.")

## 3. Merging Evictions with Boston Neighborhoods

# Create a GeoSeries containing eviction Points as its geometry, case_number as its index, and no other columns.
evictions_gdf = gpd.GeoDataFrame(evictions_df, geometry=gpd.points_from_xy(evictions_df['Longitude'], evictions_df['Latitude']))[['geometry', 'case_number']]
evictions_gdf = evictions_gdf.set_crs("EPSG:4326", allow_override=True).to_crs('EPSG:26986')

# Read shapefile containing Boston neighborhood boundaries
boston_neighborhoods_gdf = (gpd.read_file(INPUT_DATA_NEIGHBORHOODS)[['geometry', 'Name']]
                            .to_crs('EPSG:26986')
                            .rename(columns={'Name': 'neighborhood'}))

# Spatial join
evictions_neighborhoods_df = (gpd.sjoin(evictions_gdf, boston_neighborhoods_gdf, how='inner', predicate='within')
                              .drop(columns=['index_right', 'geometry'])
                              .set_index('case_number'))

## 4. Merge Evictions with Crimes

# Request computing resources.
cluster = SLURMCluster(queue='batch',
                       cores=32,
                       memory='230 GB',
                       walltime='00:20:00',
                      scheduler_options={'dashboard_address': '8787'} )
cluster.scale(jobs=1)


client = Client(cluster)

# Read crime data as Dask DataFrame, then compute back to DataFrame.
crime_df = (dd.read_csv(INPUT_DATA_CRIME + "/*.csv", dtype={'REPORTING_AREA': 'object', 'SHOOTING': 'object'})
                .dropna(subset=['Long', 'Lat', 'OCCURRED_ON_DATE'])  # Drop crimes missing latitude, longitude, or date, as they cannot be merged with panel.
                .rename(columns={'OCCURRED_ON_DATE': 'month_of_crime_incident'})
                .drop(columns=['OFFENSE_CODE_GROUP', 'OFFENSE_DESCRIPTION', 'DISTRICT', 'REPORTING_AREA', 'YEAR', 'MONTH',
                               'DAY_OF_WEEK', 'HOUR', 'UCR_PART', 'STREET', 'Location'])  # Drop unneeded columns
                .compute())
                # Must call compute here and then briefly convert back to in-memory dataset because dask_geopandas.points_from_xy not working.
crime_df.loc[:, 'INCIDENT_NUMBER'] = 1  # Replace column with 1s so we can count crimes using sum function.
# Keep track of the month of crime incident in YYYY-MM format.
crime_df.loc[:, 'month_of_crime_incident'] = pd.to_datetime(crime_df['month_of_crime_incident'].str[:10]).dt.to_period("M").astype(str)
# Convert DataFrame to GeoDataFrame.
crime_gdf = gpd.GeoDataFrame(crime_df, geometry=gpd.points_from_xy(crime_df['Long'], crime_df['Lat']))
crime_gdf = crime_gdf.set_crs("EPSG:4326", allow_override=True).to_crs("EPSG:26986")  # Convert to the correct CRS.
# Convert GeoDataFrame to Dask GeoDataFrame.
crime_dgdf = dask_geopandas.from_geopandas(crime_gdf, npartitions=N_PARTITIONS).repartition(partition_size='5 MB')


columns_for_each_outcome = []
offense_groups = ['all',
                  constants.Analysis.larceny,
                  constants.Analysis.motor_vehicle,
                  constants.Analysis.vandalism,
                  constants.Analysis.assault,
                  constants.Analysis.auto_theft,
                  constants.Analysis.larceny_motor_vehicle,
                  constants.Analysis.assault_investigate_auto_theft_vandalism]
radii = [250, 300, 350, "250_to_300", "250_to_350", "250_to_400"]
for offense_group_num, offense_group in enumerate(offense_groups):
    for radius in radii:
        
        if isinstance(radius, int):
            # Create a new GeoDataFrame with geometry equal to circles around each eviction with the current radius.
            current_evictions_gdf = evictions_gdf.copy()
            current_evictions_gdf.geometry = current_evictions_gdf.geometry.buffer(radius)
            current_evictions_dgdf = dask_geopandas.from_geopandas(current_evictions_gdf, npartitions=N_PARTITIONS).repartition(partition_size='5 MB')
        else: 
            # We need to create a "donut" around the property
            inner_radius = evictions_gdf.geometry.buffer(int(radius[:3]))
            outer_radius = evictions_gdf.geometry.buffer(int(radius[-3:]))
            donut = outer_radius.difference(inner_radius)
            current_evictions_gdf = evictions_gdf.copy()
            current_evictions_gdf.geometry = donut
            current_evictions_dgdf = dask_geopandas.from_geopandas(current_evictions_gdf, npartitions=N_PARTITIONS).repartition(partition_size='5 MB')


    
        # Merge evictions with crimes that fall into radius.
        current_evictions_crime_dgdf = dask_geopandas.sjoin(crime_dgdf, current_evictions_dgdf, how='inner', predicate='within')
        current_evictions_crime_dgdf = current_evictions_crime_dgdf.drop(columns=['geometry','index_right'])
        
        
        if offense_group != 'all':  # If we are summing a specific subcategory of crimes
            mask = current_evictions_crime_dgdf['OFFENSE_CODE'].isin(offense_group)
            # Multiply mask by 'INCIDENT_NUMBER' to zero out crimes in different offense groups.
            current_evictions_crime_dgdf['INCIDENT_NUMBER'] = current_evictions_crime_dgdf['INCIDENT_NUMBER'] * mask

                
            
        # Aggregate crimes to case-month level.
        current_panel_long = (current_evictions_crime_dgdf
                              .groupby(['case_number', 'month_of_crime_incident'])
                              .aggregate({'INCIDENT_NUMBER': 'sum'})
                              .reset_index()
                              .rename(columns={'INCIDENT_NUMBER': 'crime_incidents'})
                              .compute())

        current_panel_wide = pd.pivot(current_panel_long, index=['case_number'], columns=['month_of_crime_incident'],
                                      values='crime_incidents').reset_index().set_index('case_number')
        current_panel_wide.columns = [f'{column}_group_{offense_group_num}_crimes_{radius}m' for column in current_panel_wide.columns]
        columns_for_each_outcome.append(current_panel_wide.columns)
        value_vars_to_concat.append(dd.from_pandas(current_panel_wide, npartitions=N_PARTITIONS))

## 5. Producing the Analysis Dataset

crime_df = dd.multi.concat([dd.from_pandas(evictions_tracts_df, npartitions=N_PARTITIONS), dd.from_pandas(evictions_neighborhoods_df, npartitions=N_PARTITIONS)] + value_vars_to_concat, axis=1).compute()
# For evictions not matched to any crimes, fill NA values with 0.
for group_of_columns in columns_for_each_outcome:
    crime_df[group_of_columns] = crime_df[group_of_columns].fillna(0)

# Generate a variable indicating judgment in favor of defendant.
crime_df.loc[:, 'judgment_for_defendant'] = 0
defendant_won_mask = ((crime_df['disposition_found'] == "Dismissed") |
                      (crime_df['judgment_for_pdu'] == "Defendant") |
                      (crime_df['disposition'].str.contains('R 41(a)(1) Voluntary Dismissal', regex=False)))
crime_df.loc[defendant_won_mask, 'judgment_for_defendant'] = 1

# Generate a variable indicating judgement in favor of plaintiff.
crime_df.loc[:, 'judgment_for_plaintiff'] = 1 - crime_df['judgment_for_defendant']




# Create variables containing pre-treatment outcomes
outcomes = [f"group_{group_num}_crimes_{radius}m" for group_num in range(len(offense_groups)) for radius in radii]

for outcome in outcomes:
    # Save value variable names and dictionary mapping months to integers
    triplet = get_value_variable_names(crime_df, outcome)
    weekly_value_vars_crime, month_to_int_dictionary, _ = triplet
    
    # Create variable equal to sum of crimes in 2017
    columns_from_2017 = [column for column in crime_df.columns if column.startswith('2017') and column.endswith(outcome)]
    crime_df.loc[:, f'total_twenty_seventeen_{outcome}'] = crime_df[columns_from_2017].sum(axis=1)
    
    # Create variable equal to sum of crimes in 2018 
    columns_from_2019 = [column for column in crime_df.columns if column.startswith('2018') and column.endswith(outcome)]
    crime_df.loc[:, f'total_twenty_eighteen_{outcome}'] = crime_df[columns_from_2019].sum(axis=1)
    
    # Create variable equal to change in the total number of crimes between 2019 and 2017
    crime_df.loc[:, f'pre_treatment_change_in_{outcome}'] = crime_df[f'total_twenty_eighteen_{outcome}'] - crime_df[f'total_twenty_seventeen_{outcome}']
    
    # Create variable equal to change in the total number of crimes in the two years prior to treatment
    crime_df_copy = crime_df.copy().reset_index()
    crime_df_copy = pd.melt(crime_df_copy,
                            id_vars=['case_number',
                                     'latest_docket_month'],
                            value_vars=weekly_value_vars_crime,
                            var_name='month')
    crime_df_copy.loc[:, 'month'] = crime_df_copy['month'].str[:7]
    crime_df_copy.loc[:, ['latest_docket_month', 'month']] = crime_df_copy[['latest_docket_month', 'month']].replace(month_to_int_dictionary)
    crime_df_copy.loc[:, 'treatment_relative_month'] = crime_df_copy['month'] - crime_df_copy['latest_docket_month']
    crime_df_copy = crime_df_copy.loc[crime_df_copy['treatment_relative_month'].isin([-12, 0]), ['case_number', 'treatment_relative_month', 'value']]
    crime_df_copy = crime_df_copy.pivot(index='case_number', columns='treatment_relative_month', values='value')
    relative_pre_treatment_change_in_outcome = (crime_df_copy[0] - crime_df_copy[-12]).rename(f'relative_pre_treatment_change_in_{outcome}')
    crime_df = pd.concat([crime_df, relative_pre_treatment_change_in_outcome], axis=1)
    

for_cause_mask = crime_df['initiating_action'].isin(["SP Summons and Complaint - Cause",
                                                     "Summary Process - Residential-Cause other than Non payment of rent.",
                                                     "SP Transfer - Cause"])
crime_df.loc[:, 'for_cause'] = np.where(for_cause_mask, 1, 0)

no_cause_mask = crime_df['initiating_action'].isin(["SP Summons and Complaint - No Cause",
                                                     "SP Transfer- No Cause"])
crime_df.loc[:, 'no_cause'] = np.where(no_cause_mask, 1, 0)

non_payment_of_rent_mask = crime_df['initiating_action'].isin(["SP Summons and Complaint - Non-payment of Rent",
                                                               "SP Transfer - Non-payment of Rent"])
crime_df.loc[:, 'non_payment'] = np.where(non_payment_of_rent_mask, 1, 0)

# Create case resolution variables
panel_E_columns = ['dismissed', 'defaulted', 'heard']
origin_columns = ['disposition_found', 'disposition_found',
                  'disposition_found']
target_values = ["Dismissed", "Defaulted", "Heard"]

for dummy_column, origin_column, target_value in zip(panel_E_columns, origin_columns, target_values):
    crime_df.loc[:, dummy_column] = np.where(crime_df[origin_column] == target_value, 1, 0)

crime_df.to_csv(OUTPUT_DATA_CRIME)

#####
